Saving Datapackage
Adjusted Naming of foreign_keys to be the same as in tabular, as they are named "foreignKeys"
Changed get_foreign_keys, the bus names where actually the entries that would go into "bus.csv" and not the correct fields.

For Building the Package datapackage.Package is used to infer datatypes.
1. Save elements and sequences with defined Naming
2. Build Package
3. Add Foreign keys to descriptors
4. Use updated descriptors to create new Package and save that package as json.

Then Foreign keys are added and the